Ensure that the H2Proxy request has schema in the URI
Currently, H2Proxy request builder is creating a full request URI by combining base URI and a request path. A base URI, however might not have scheme provided and when such a request is sent to certain HTTP servers they will treat this URL as path and mishandle the request. This commit checks if base URI has a scheme and prepends http:// if not.