Ensure that grpc-status code is properly propagated in the thrown exception
Originally, the Status instance that is created in response of the request exception always had its code set to INTERNAL (13). This prevents clients from properly understanding what the server is trying to tell them. This commit reads the code from grpc-status header (or rather trailer) field and uses it as a code for an exception to be thrown, and only falls back onto UNKNOWN (2) if the grpc-status is missing.