Ensure that base URI contains scheme
Currently the H2 config class accepts base URIs that contain no schema. When a request is sent through proxy, certain services expecting fully qualified URI that include a scheme and fail if a scheme isn't present. This commit introduces validation of a base URI at the time of config instantiation and throws an exception if scheme is not provided.