Change Client::$config property back to protected
Previous refactoring #62, changed visibility of the `Client::$config` variable from `protected` to `private` as it isn't used anywhere outside of the class in this library. However, I have later discovered that we extend the `Client` class in the `grphp-balancer` library and it uses the `$config` property from the parent. Relax the visibility to fix this problem.